UI Update
I changed the title screen setup and added scripts for the translator to hopefully function. I added new objects for journal and translator and a sign to test the translator but haven't tested it yet. I added a controls icon to be displayed while playing for the player to access for controls info but haven't tested that either.